dev-requirements.txt
flake8>=3.7,<4.0
pre-commit>=2.2,<2.3
ipdb>=0.13,<0.14
coverage<5.0.0
python-coveralls>=2.9,<3.0
django-extensions>=2.2,<2.3
selenium==3.141.0
sendgrid==6.4.7
